+++ src/components/index.js
@@ -94,7 +94,7 @@ export EditDescriptionTile from '@plone/volto/components/manage/Tiles/Descriptio
 export EditTitleTile from '@plone/volto/components/manage/Tiles/Title/Edit';
 export EditTextTile from '@plone/volto/components/manage/Tiles/Text/Edit';
 export EditImageTile from '@plone/volto/components/manage/Tiles/Image/Edit';
-export EditSummaryBoxTile from '@plone/volto/components/manage/Tiles/SummaryBox/Edit';
+export EditProxyTile from '@plone/volto/components/manage/Tiles/Proxy/Edit';
 export EditVideoTile from '@plone/volto/components/manage/Tiles/Video/Edit';
 export EditTile from '@plone/volto/components/manage/Tiles/Tile/Edit';
 export EditHeroImageLeftTile from '@plone/volto/components/manage/Tiles/HeroImageLeft/Edit';
@@ -106,7 +106,7 @@ export ViewDescriptionTile from '@plone/volto/components/manage/Tiles/Descriptio
 export ViewTitleTile from '@plone/volto/components/manage/Tiles/Title/View';
 export ViewTextTile from '@plone/volto/components/manage/Tiles/Text/View';
 export ViewImageTile from '@plone/volto/components/manage/Tiles/Image/View';
-export ViewSummaryBoxTile from '@plone/volto/components/manage/Tiles/SummaryBox/View';
+export ViewProxyTile from '@plone/volto/components/manage/Tiles/Proxy/View';
 export ViewVideoTile from '@plone/volto/components/manage/Tiles/Video/View';
 export ViewMapTile from '@plone/volto/components/manage/Tiles/Maps/View';
 export ViewHTMLTile from '@plone/volto/components/manage/Tiles/HTML/View';